Week 11 Commit 2
Sound System Refinement
Level Expansion
Health Component
Enemy Attack

+++ Assets/Scripts/HealthBar.cs
@@ -11,7 +11,7 @@ public class HealthBar : MonoBehaviour
 		m_TextMeshPro = GetComponentInChildren<TMPro.TextMeshProUGUI>();
 	}
 
-	public void SetText(int amount)
+	public void SetText(float amount)
 	{
 		m_TextMeshPro.text = amount.ToString();
 	}
